# docs/swagger.yaml
openapi: 3.0.3
info:
  title: AlpineLink 雪域 API
  description: |
    AlpineLink 雪域高端登山户外社交平台后端 API 文档。

    ## 认证方式
    大多数接口需要在请求头中携带 JWT Token：
    ```
    Authorization: Bearer <token>
    ```
    Token 通过 `/api/auth/login` 获取。

    ## 接口分级
    - **P0**：核心功能，认证模块
    - **P1**：主要业务，帖子、评论、山峰/路线、活动、消息
    - **P2**：扩展功能，向导、俱乐部、SOS、后台统计

    ## 统一响应格式
    ```json
    { "code": 0, "message": "success", "data": {} }
    ```

    ## WebSocket 实时消息
    连接地址：`ws://<host>/ws?token=<jwt_token>`

    消息格式：
    ```json
    { "type": "message", "chatId": 1, "content": "你好", "msgType": "text" }
    ```

    服务端推送格式：
    ```json
    { "type": "message", "data": { "id": 1, "chatId": 1, "senderId": 2, "content": "你好", "msgType": "text", "createdAt": "2026-01-01T00:00:00Z" } }
    ```
  version: 1.0.0
  contact:
    name: AlpineLink CEO
    url: https://desktop-tutorial-production-182a.up.railway.app
servers:
  - url: https://desktop-tutorial-production-182a.up.railway.app
    description: Railway 生产环境
  - url: http://localhost:3000
    description: 本地开发环境
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Token，通过 `/api/auth/login` 获取
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          example: 1003
        message:
          type: string
          example: 未登录或Token无效
        data:
          nullable: true
          example: null
...
    Route:
      type: object
      properties:
        id:
          type: integer
          example: 1
        peak_id:
          type: integer
          example: 1
        name:
          type: string
          example: 标准南坡路线
        difficulty:
          type: string
          enum: [较易, 中等, 难, 极难]
          example: 极难
        description:
          type: string
          example: 从尼泊尔侧登顶珠峰的经典路线
        cover_image:
          type: string
          format: uri
          example: https://example.com/route.jpg
        altitude:
          type: integer
          example: 8848
        duration_days:
          type: integer
          example: 60
        best_season:
          type: string
          example: 4月-5月
        region:
          type: string
          example: 尼泊尔
        status:
          type: string
          example: active

    # ────────── 活动 ──────────
...
    Badge:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 初登者
        description:
          type: string
          example: 完成第一次登山记录
        icon:
          type: string
          format: uri
          nullable: true
          example: https://example.com/badge_icon.png
        category:
          type: string
          example: 登山
        condition_type:
          type: string
          example: post_count
        condition_value:
          type: integer
          example: 1
        unlocked_at:
          type: string
          format: date-time
          nullable: true
          example: null
        progress:
          type: integer
          nullable: true
          example: 0

    # ────────── 装备 ──────────
...
  parameters:
    PageParam:
      name: page
      in: query
      description: 页码，从 1 开始
      schema:
        type: integer
        default: 1
        minimum: 1
    LimitParam:
      name: limit
      in: query
      description: 每页条数
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100

  # ────────── 通用响应 ──────────
  responses:
    Unauthorized:
      description: 未登录或 Token 无效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 1003
            message: 未登录或Token无效
            data: null
    Forbidden:
      description: 权限不足
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 1004
            message: 权限不足
            data: null
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 1005
            message: 资源不存在
            data: null
...
paths:
...
  /api/routes:
    get:
      tags: [Routes]
      summary: 路线列表
      description: 获取所有路线列表，支持分页和按山峰筛选。
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: peak_id
          in: query
          description: 按山峰 ID 筛选
          schema:
            type: integer
        - name: difficulty
          in: query
          description: 按难度筛选
          schema:
            type: string
            enum: [较易, 中等, 难, 极难]
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          list:
                            type: array
                            items:
                              $ref: '#/components/schemas/Route'
                          total:
                            type: integer
              example:
                code: 0
                message: success
                data:
                  list:
                    - id: 1
                      peak_id: 1
                      name: 标准南坡路线
                      difficulty: 极难
                      altitude: 8848
                      duration_days: 60
                  total: 4

    post:
      tags: [Routes]
      summary: 创建路线（需 admin 权限）
      description: 创建新的攀登路线，仅管理员可操作。
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [peak_id, name, difficulty]
              properties:
                peak_id:
                  type: integer
                  example: 1
                name:
                  type: string
                  example: 北坡路线
                difficulty:
                  type: string
                  enum: [较易, 中等, 难, 极难]
                  example: 极难
                description:
                  type: string
                  example: 从西藏侧登顶的经典路线
                cover_image:
                  type: string
                  format: uri
                  example: https://example.com/route2.jpg
                altitude:
                  type: integer
                  example: 8848
                duration_days:
                  type: integer
                  example: 55
                best_season:
                  type: string
                  example: 5月
                region:
                  type: string
                  example: 西藏
            example:
              peak_id: 1
              name: 北坡路线
              difficulty: 极难
              description: 从西藏侧登顶珠峰
              altitude: 8848
              duration_days: 55
              best_season: 5月
              region: 西藏
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Route'
              example:
                code: 0
                message: 创建成功
                data:
                  id: 5
                  peak_id: 1
                  name: 北坡路线
                  difficulty: 极难
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/routes/{id}:
    get:
      tags: [Routes]
      summary: 路线详情
      description: 获取指定 ID 的路线详情。
      parameters:
        - name: id
          in: path
          required: true
          description: 路线 ID
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Route'
              example:
                code: 0
                message: success
                data:
                  id: 1
                  peak_id: 1
                  name: 标准南坡路线
                  difficulty: 极难
                  description: 从尼泊尔侧登顶珠峰的经典路线
                  altitude: 8848
                  duration_days: 60
                  best_season: 4月-5月
                  region: 尼泊尔
                  status: active
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Routes]
      summary: 更新路线（需 admin 权限）
      description: 更新指定 ID 的路线信息，仅管理员可操作。
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 路线 ID
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                difficulty:
                  type: string
                  enum: [较易, 中等, 难, 极难]
                description:
                  type: string
                cover_image:
                  type: string
                  format: uri
                altitude:
                  type: integer
                duration_days:
                  type: integer
                best_season:
                  type: string
                region:
                  type: string
                status:
                  type: string
            example:
              description: 更新后的路线描述
              best_season: 4月-6月
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Route'
              example:
                code: 0
                message: 更新成功
                data:
                  id: 1
                  name: 标准南坡路线
                  best_season: 4月-6月
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ────────── 4.5 活动/组队模块 ──────────
...
  /api/badges/my:
    get:
      tags: [Badges]
      summary: 获取当前用户的徽章解锁状态（需登录）
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 徽章列表（含解锁进度）
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Badge'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/badges/check:
    post:
      tags: [Badges]
      summary: 检查并发放新解锁的徽章（需登录）
      description: 依据当前用户的帖子数、登顶海拔、点赞数、粉丝数重新计算可解锁徽章，并写入数据库。
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 本次新解锁的徽章列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  unlocked:
                    type: array
                    items:
                      $ref: '#/components/schemas/Badge'
                  count:
                    type: integer
                    example: 2
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ────────── 5.9 装备租借模块 ──────────
...